requirements.txt
torch>=2.0
torchvision
numpy
pillow
tqdm
cleanfid
torch-fidelity
opencv-python
pandas
scipy
